app/requirements.txt
fastapi
uvicorn
pytesseract
pillow
ocrmypdf
python-multipart
pdfminer.six